feat: Added thread pool and fixed framebuffer issues!

+++ Engine/include/core/UniquelyIdentifiable.hpp
@@ -5,7 +5,7 @@
 
 namespace Disarray {
 
-	using Identifier = std::size_t;
+	using Identifier = std::uint64_t;
 
 	template <class Derived> struct UniquelyIdentifiable {
 		template <class T = Derived> Identifier hash() { return static_cast<T&>(*this).hash_impl(); }
